+++ packages/admin/admin-rte/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @comet/admin-rte
 
+## 8.9.0
+
+### Patch Changes
+
+- @comet/admin@8.9.0
+- @comet/admin-icons@8.9.0
+
+## 8.8.0
+
+### Patch Changes
+
+- @comet/admin@8.8.0
+- @comet/admin-icons@8.8.0
+
 ## 8.7.1
 
 ### Patch Changes