+++ packages/admin/admin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/admin
 
+## 8.9.0
+
+### Patch Changes
+
+- @comet/admin-icons@8.9.0
+
+## 8.8.0
+
+### Patch Changes
+
+- @comet/admin-icons@8.8.0
+
 ## 8.7.1
 
 ### Patch Changes